statistic: refactor statistics
