+++ statistics/handle/cache/BUILD.bazel
@@ -15,9 +15,9 @@ go_library(
         "//sessionctx",
         "//sessionctx/variable",
         "//statistics",
-        "//statistics/handle/internal/cache",
-        "//statistics/handle/internal/cache/lru",
-        "//statistics/handle/internal/cache/mapcache",
+        "//statistics/handle/cache/internal",
+        "//statistics/handle/cache/internal/lru",
+        "//statistics/handle/cache/internal/mapcache",
         "//types",
         "//util/chunk",
         "//util/memory",
