+++ statistics/handle/cache/internal/inner.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cache
+package internal
 
 import (
 	"github.com/pingcap/tidb/statistics"
