+++ statistics/handle/cache/internal/mapcache/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "mapcache",
     srcs = ["map_cache.go"],
-    importpath = "github.com/pingcap/tidb/statistics/handle/internal/cache/mapcache",
+    importpath = "github.com/pingcap/tidb/statistics/handle/cache/internal/mapcache",
     visibility = ["//statistics/handle:__subpackages__"],
     deps = [
         "//statistics",
-        "//statistics/handle/internal/cache",
+        "//statistics/handle/cache/internal",
     ],
 )
